refactor:Refactoring exports schemas and ctr. Move joi schemas to dir/schemas
