+++ models/user.js
@@ -37,30 +37,6 @@ userSchema.post("save", handleMongooseError);
 
 const User = model("user", userSchema);
 
-const registerSchema = Joi.object({
-  name: Joi.string().min(2).required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const loginSchema = Joi.object({
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const updateSchema = Joi.object({
-  name: Joi.string().min(2),
-  email: Joi.string().pattern(emailRegexp),
-  password: Joi.string().min(6),
-});
-
-const schemas = {
-  registerSchema,
-  loginSchema,
-  updateSchema,
-};
-
 module.exports = {
   User,
-  schemas,
 };
