+++ routes/api/users.js
@@ -1,33 +1,41 @@
 const express = require("express");
 const router = express.Router();
-const usersController = require("../../controllers/users");
+const {
+  updateTheme,
+  helpRequest,
+  currentUser,
+  updateUser,
+} = require("../../controllers/users");
 const { multerUpload, validateBody, auth } = require("../../middlewares");
-const { themeSchema, helpRequestSchema } = require("../../schemas");
-const { schemas } = require("../../models/user");
+const {
+  themeSchema,
+  helpRequestSchema,
+  updateSchema,
+} = require("../../schemas");
 const { controllerWrapper } = require("../../helpers");
 
 router.patch(
   "/themes",
   auth,
   validateBody(themeSchema),
-  controllerWrapper(usersController.updateTheme)
+  controllerWrapper(updateTheme)
 );
 
 router.patch(
   "/help",
   auth,
   validateBody(helpRequestSchema),
-  controllerWrapper(usersController.helpRequest)
+  controllerWrapper(helpRequest)
 );
 
-router.get("/current", auth, controllerWrapper(usersController.currentUser));
+router.get("/current", auth, controllerWrapper(currentUser));
 
 router.patch(
   "/",
   auth,
   multerUpload.single("newAvatar"),
-  validateBody(schemas.updateSchema),
-  usersController.updateUser
+  validateBody(updateSchema),
+  controllerWrapper(updateUser)
 );
 
 module.exports = router;
