+++ schemas/auth.js
@@ -1,0 +1,16 @@
+const Joi = require("joi");
+
+const emailRegexp = /^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$/;
+
+const registerSchema = Joi.object({
+  name: Joi.string().min(2).required(),
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+});
+
+const loginSchema = Joi.object({
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+});
+
+module.exports = { registerSchema, loginSchema };
